add an intermediate humanoid controller (stab but no torque check)
