+++ etc/icub/arm.yaml
@@ -1,0 +1,6 @@
+BEHAVIOR:
+  name : generic::cartesian
+  task_names: [lh]
+  relative_targets: [[0, 0, 0.1]]
+  trajectory_duration: 1
+  loop: true
