+++ etc/icub/clapping.yaml
@@ -1,0 +1,4 @@
+BEHAVIOR:
+  name : humanoid::clapping
+  trajectory_duration : 1.0
+  motion_size : 1.1
