+++ etc/icub/collision_thresholds.yaml
@@ -1,0 +1,25 @@
+# torque discrepancy threshold between tsid and sensors
+# used to detect collisions
+
+leg_left_1_joint: 1e+10
+leg_left_2_joint: 1e+10
+leg_left_3_joint: 1e+10
+leg_left_4_joint: 1e+10
+leg_left_5_joint: 1e+10
+leg_left_6_joint: 1e+10
+leg_right_1_joint: 1e+10
+leg_right_2_joint: 1e+10
+leg_right_3_joint: 1e+10
+leg_right_4_joint: 1e+10
+leg_right_5_joint: 1e+10
+leg_right_6_joint: 1e+10
+torso_1_joint: 1e+00
+torso_2_joint: 1e+00
+arm_left_1_joint: 1e+00
+arm_left_2_joint: 1e+00
+arm_left_3_joint: 1e+00
+arm_left_4_joint: 1e+00
+arm_right_1_joint: 1e+00
+arm_right_2_joint: 1e+00
+arm_right_3_joint: 1e+00
+arm_right_4_joint: 1e+00
