+++ etc/icub/squat.yaml
@@ -1,0 +1,4 @@
+BEHAVIOR:
+  name: humanoid::squat
+  trajectory_duration: 1.0
+  motion_size: 0.1
