+++ etc/icub/stab_double_support.yaml
@@ -1,0 +1,9 @@
+filter_size: 7
+torso_max_roll: 0.25
+com: [0.001, 0.001, 0.0003, 0.0003, 0.00000003, 0.00000003]
+ankle: [0.25, 0.25, 0.0001, 0.001, 0.00001, 0.00001]
+ffda: [0.0003, 0.0, 0]
+activate_zmp: true
+zmp_w: [1, 1, 1, 2, 2, 2]
+zmp_d: [0, 0, 0.0005, 0.0001, 0.002, 0]
+zmp_p: [0, 0, 2, 10.0, 10.0, 0]
