+++ etc/icub/stab_fixed_base.yaml
@@ -1,0 +1,9 @@
+filter_size: 7
+torso_max_roll: 0.25
+com: [0, 0, 0, 0, 0, 0]
+ankle: [0, 0, 0, 0, 0, 0]
+ffda: [0, 0, 0]
+activate_zmp: false
+zmp_w: [0, 0, 0, 0, 0, 0]
+zmp_d: [0, 0, 0, 0, 0, 0]
+zmp_p: [0, 0, 0, 0, 0, 0]
