+++ etc/icub/traj_teleop1.yaml
@@ -1,0 +1,5 @@
+BEHAVIOR:
+  name : generic::cartesian_traj
+  trajectories: traj_teleop1/trajectory.yaml
+  loop: true
+  scale: 0.5
