+++ etc/icub/traj_teleop1/trajectory.yaml
@@ -1,0 +1,4 @@
+refs:
+  lh: lh.csv
+  rh: rh.csv
+  com: com.csv
