+++ etc/icub/walk_on_spot.yaml
@@ -1,0 +1,7 @@
+BEHAVIOR:
+  name: humanoid::walk-on-spot
+  traj_com_duration: 1.0
+  traj_foot_duration: 0.5
+  step_height: 0.05
+  stop_duration: 0.1
+  stop_height: 0.0015
