+++ include/inria_wbc/controllers/pos_tracker.hpp
@@ -7,6 +7,7 @@
 namespace inria_wbc {
     namespace controllers {
       
+        // generic position tracker, for both fixed-base and floatin-base robots
         class PosTracker : public Controller {
         public:
             PosTracker(const YAML::Node& config);
